Bid -> add error management

+++ src/main/java/com/nnk/poseidoninc/Service/BidService.java
@@ -1,5 +1,6 @@
 package com.nnk.poseidoninc.Service;
 
+import com.nnk.poseidoninc.Exeption.BidNotFoundException;
 import com.nnk.poseidoninc.Model.Bid;
 import com.nnk.poseidoninc.Model.Dto.BidDto;
 import com.nnk.poseidoninc.Repository.BidRepository;
@@ -23,13 +24,13 @@ public class BidService {
     }
 
 
-    public List<BidDto> findAll(){
+    public List<BidDto> findAll() {
 
         Iterable<Bid> bids = bidRepository.findAll();
 
         List<BidDto> bidList = new ArrayList<>();
 
-        for (Bid bid : bids){
+        for (Bid bid : bids) {
             bidList.add(convertBidToBidDto(bid));
         }
 
@@ -43,14 +44,26 @@ public class BidService {
     }
 
     public BidDto findById(int id) {
-        Bid bid = bidRepository.findById(id).get();
+        Optional<Bid> bid = bidRepository.findById(id);
 
-        return convertBidToBidDto(bid);
+        if (bid.isEmpty()) {
+            throw new BidNotFoundException();
+        }
+
+        return convertBidToBidDto(bid.get());
+
+
     }
 
     public BidDto update(BidDto bidDto) {
 
-        Bid bid = bidRepository.findById(bidDto.getBidId()).get();
+        Optional<Bid> optionalBid = bidRepository.findById(bidDto.getBidId());
+
+        if (optionalBid.isEmpty()) {
+            throw new BidNotFoundException();
+        }
+
+        Bid bid = optionalBid.get();
         bid.setAccount(bidDto.getAccount());
         bid.setType(bidDto.getType());
         bid.setBidQuantity(bidDto.getBidQuantity());
@@ -59,6 +72,10 @@ public class BidService {
     }
 
     public void delete(int id) {
+
+        if (bidRepository.findById(id).isEmpty()) {
+            throw new BidNotFoundException();
+        }
 
         bidRepository.deleteById(id);
     }
